pyproject.toml
[project]
name = "lab-complaint"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cryptography>=46.0.3",
    "databases>=0.9.0",
    "fastapi[standard]>=0.127.0",
    "jinja2>=3.1.6",
    "mysql-connector-python>=9.5.0",
    "pydantic>=2.12.5",
    "pymysql>=1.1.2",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.40.0",
]
